Extract on file per test, one test per metric
